Type properties
Add DBMS Type description classes (implements PSR Container interface)

Change StatementBuilder method to get DBMS type from column structure property

WIP PostgreSQL

+++ tests/reference/PDO_SQLiteCreateInsertUpdateDelete_create.sql
@@ -2,6 +2,6 @@ CREATE TABLE [Employees] (
   [id] INTEGER NOT NULL, 
   [name] TEXT, 
   [gender] TEXT, 
-  [salary] REAL(7), 
+  [salary] REAL, 
   CONSTRAINT [pk_id] PRIMARY KEY ([id])
 )
